update logging so pwm signal is received as a uint32

+++ libraries/DataFlash/LogFile.cpp
@@ -1810,7 +1810,7 @@ void DataFlash_Class::Log_Write_RPM(const AP_RPM &rpm_sensor)
     WriteBlock(&pkt, sizeof(pkt));
 }
 
-void DataFlash_Class::Log_Write_QX1Gimbal(float ref, float rout,float pout, uint16_t rpwm, uint16_t ppwm)
+void DataFlash_Class::Log_Write_QX1Gimbal(float ref, float rout,float pout, uint32_t rpwm, uint32_t ppwm)
 {
     struct log_QX1Gimbal pkt = {
         LOG_PACKET_HEADER_INIT(LOG_QX1GIMBAL_MSG),
